refactor(forms): filename changed config -> makeConfigurableComponent
CACTUS-795
